refactor: better scope eta messages const
