+++ src/handlers/event/eta.rs
@@ -1,11 +1,29 @@
-use crate::{consts, utils};
-
 use color_eyre::eyre::Result;
 use once_cell::sync::Lazy;
 use poise::serenity_prelude::{Context, Message};
+use rand::seq::SliceRandom;
 use regex::Regex;
 
 static ETA_REGEX: Lazy<Regex> = Lazy::new(|| Regex::new(r"\beta\b").unwrap());
+
+pub const ETA_MESSAGES: [&str; 16] = [
+	"Sometime",
+	"Some day",
+	"Not far",
+	"The future",
+	"Never",
+	"Perhaps tomorrow?",
+	"There are no ETAs",
+	"No",
+	"Nah",
+	"Yes",
+	"Yas",
+	"Next month",
+	"Next year",
+	"Next week",
+	"In Prism Launcher 2.0.0",
+	"At the appropriate juncture, in due course, in the fullness of time",
+];
 
 pub async fn handle(ctx: &Context, message: &Message) -> Result<()> {
 	if !ETA_REGEX.is_match(&message.content) {
@@ -14,7 +32,9 @@ pub async fn handle(ctx: &Context, message: &Message) -> Result<()> {
 
 	let response = format!(
 		"{} <:pofat:1031701005559144458>",
-		utils::random_choice(consts::ETA_MESSAGES)?
+		ETA_MESSAGES
+			.choose(&mut rand::thread_rng())
+			.unwrap_or(&"sometime")
 	);
 
 	message.reply(ctx, response).await?;
